+++ Events.cs
@@ -9,10 +9,12 @@ public class Events
     /// Triggered once the server has loaded and VAMP.Core has initialized.
     /// </summary>
     public static Action OnCoreLoaded;
+
     /// <summary>
-    /// Triggered if the server has been wiped. 
-    /// This can be used to clear any data that should not persist between server restarts.
-    /// NOT READY FOR USE
+    /// Triggered when the server has been wiped.
+    /// This event indicates that server data has been cleared and provides information about the wipe status.
+    /// If the boolean parameter is true, files in the `_WipeData` folder were automatically deleted by VAMP.
+    /// This event is always invoked before OnCoreLoaded.
     /// </summary>
-    public static Action OnServerWiped;
+    public static Action<bool> OnServerWiped;
 }